+++ deep_causality_tensor/src/types/causal_tensor/op_tensor_scalar/mod.rs
@@ -12,14 +12,14 @@ use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign};
 // Implementation for `&CausalTensor<T> + T`
 impl<T> Add<T> for &CausalTensor<T>
 where
-    T: Add<T, Output = T> + Copy,
+    T: Add<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
 
     fn add(self, rhs: T) -> Self::Output {
         let mut new_data = self.data.clone();
         for item in &mut new_data {
-            *item = (*item).add(rhs);
+            *item = item.clone().add(rhs.clone());
         }
         CausalTensor {
             data: new_data,
@@ -32,7 +32,7 @@ where
 // Implementation for `CausalTensor<T> + T` (consuming)
 impl<T> Add<T> for CausalTensor<T>
 where
-    T: Add<T, Output = T> + Copy,
+    T: Add<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
     fn add(self, rhs: T) -> Self::Output {
@@ -43,11 +43,11 @@ where
 // Implementation for `CausalTensor<T> += T` (in-place)
 impl<T> AddAssign<T> for CausalTensor<T>
 where
-    T: AddAssign + Copy,
+    T: AddAssign + Clone,
 {
     fn add_assign(&mut self, rhs: T) {
         for item in &mut self.data {
-            item.add_assign(rhs);
+            item.add_assign(rhs.clone());
         }
     }
 }
@@ -57,14 +57,14 @@ where
 // Implementation for `&CausalTensor<T> - T`
 impl<T> Sub<T> for &CausalTensor<T>
 where
-    T: Sub<T, Output = T> + Copy,
+    T: Sub<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
 
     fn sub(self, rhs: T) -> Self::Output {
         let mut new_data = self.data.clone();
         for item in &mut new_data {
-            *item = (*item).sub(rhs);
+            *item = item.clone().sub(rhs.clone());
         }
         CausalTensor {
             data: new_data,
@@ -77,7 +77,7 @@ where
 // Implementation for `CausalTensor<T> - T` (consuming)
 impl<T> Sub<T> for CausalTensor<T>
 where
-    T: Sub<T, Output = T> + Copy,
+    T: Sub<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
     fn sub(self, rhs: T) -> Self::Output {
@@ -88,11 +88,11 @@ where
 // Implementation for `CausalTensor<T> -= T` (in-place)
 impl<T> SubAssign<T> for CausalTensor<T>
 where
-    T: SubAssign + Copy,
+    T: SubAssign + Clone,
 {
     fn sub_assign(&mut self, rhs: T) {
         for item in &mut self.data {
-            item.sub_assign(rhs);
+            item.sub_assign(rhs.clone());
         }
     }
 }
@@ -102,14 +102,14 @@ where
 // Implementation for `&CausalTensor<T> * T`
 impl<T> Mul<T> for &CausalTensor<T>
 where
-    T: Mul<T, Output = T> + Copy,
+    T: Mul<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
 
     fn mul(self, rhs: T) -> Self::Output {
         let mut new_data = self.data.clone();
         for item in &mut new_data {
-            *item = (*item).mul(rhs);
+            *item = item.clone().mul(rhs.clone());
         }
         CausalTensor {
             data: new_data,
@@ -122,7 +122,7 @@ where
 // Implementation for `CausalTensor<T> * T` (consuming)
 impl<T> Mul<T> for CausalTensor<T>
 where
-    T: Mul<T, Output = T> + Copy,
+    T: Mul<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
     fn mul(self, rhs: T) -> Self::Output {
@@ -133,11 +133,11 @@ where
 // Implementation for `CausalTensor<T> *= T` (in-place)
 impl<T> MulAssign<T> for CausalTensor<T>
 where
-    T: MulAssign + Copy,
+    T: MulAssign + Clone,
 {
     fn mul_assign(&mut self, rhs: T) {
         for item in &mut self.data {
-            item.mul_assign(rhs);
+            item.mul_assign(rhs.clone());
         }
     }
 }
@@ -147,14 +147,14 @@ where
 // Implementation for `&CausalTensor<T> / T`
 impl<T> Div<T> for &CausalTensor<T>
 where
-    T: Div<T, Output = T> + Copy,
+    T: Div<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
 
     fn div(self, rhs: T) -> Self::Output {
         let mut new_data = self.data.clone();
         for item in &mut new_data {
-            *item = (*item).div(rhs);
+            *item = item.clone().div(rhs.clone());
         }
         CausalTensor {
             data: new_data,
@@ -167,7 +167,7 @@ where
 // Implementation for `CausalTensor<T> / T` (consuming)
 impl<T> Div<T> for CausalTensor<T>
 where
-    T: Div<T, Output = T> + Copy,
+    T: Div<T, Output = T> + Clone,
 {
     type Output = CausalTensor<T>;
     fn div(self, rhs: T) -> Self::Output {
@@ -178,11 +178,11 @@ where
 // Implementation for `CausalTensor<T> /= T` (in-place)
 impl<T> DivAssign<T> for CausalTensor<T>
 where
-    T: DivAssign + Copy,
+    T: DivAssign + Clone,
 {
     fn div_assign(&mut self, rhs: T) {
         for item in &mut self.data {
-            item.div_assign(rhs);
+            item.div_assign(rhs.clone());
         }
     }
 }
